[MassStorageDevice] Added a disk image
It seems that we can't handle even relatively small (10MB) disk images
with Umap2, as in many cases the host gets a read timeout when the host
tries to read multiple (>0x80) blocks from the device, and then the host
resets the device. So we need a smaller image to successfully mount our
image. I found that a 3M FAT32 image will do the job (at least on my
linux machine).
To simplify the usage of the library, I added a small (3MB) compressed disk
image to the repository (can be found in the "data" folder).
As a compressed file it takes only 3KB.
Also, a more helpful message is not printed if there is not stick.img.