More robust support for USBVendor and USBClass
Some drivers really misbehave in some cases, messing up the addressing
for the USBVendor/USBClass.
We deal with them by spreading those whenever we can between the
USBActors.
There's still some work to do, especially when the Vendor/Class are used
with interface addressing before setting the configuration, but we'll
get to that.