[USBDevice] trim GET_DESCRIPTOR responses according to wLength
This is a revert to commit 0dd8f031e966ef41d566d442afd9e3d477067de1. At
least one descriptor (configuration) is requested twice during
enumeration by all hosts that we know of, but there are some hosts that
request partial descriptors (device, string) during enumeration (we've
observed that on OSX el-capitan), so instead of exception only for
configuration descriptors, we now trim all responses to GET_DESCRIPTOR
requests according to the wLength field in the setup request.

This might cause issues in fuzzing, but to our understanding, this
issues are minimal.