[Apps] Changed phy callback API
Since some Phys stop getting requests after a while, the name of the
packet_processed callback was changed to should_stop_phy, and instead of
being called only after an EP0 packet received (in GadgetFsPhy) it
should be called in each iteration of the run loop in the Phy.

Note that this means that applications should not rely on the call count
to this callback, only on other triggers and timers.