[GadgetFsPhy] cache opened endpoint fds
If soem cases, gadgetfs disconnects, reconnects and then reconfigures
the USB device. In those cases, we did not close the file descriptors
for the endpoints that were configured before. However, closing them
without closing the control endpoint proved problematic, so instead, we
cache the opened endpoint file descriptors and don't close them until we
stop the phy. In case of reconfiguration, we only update the information
regarding the updated endpoints and leave the rest as they are.