[fuzzing] can now call mutable on bound methods as well
Dirty hack to deal with cases of functions that are received as *bound methods*
When we use mutable as a decorator:

    @mutable('some stage')
    def my_stage(self, inputs):

We can define wrapper as:

    def wrapper(self, *args, **kwargs):

However, if we call mutable explicitly on a bound method:

    mutable(self.my_stage)

'wrapper' will have no reference to 'self', and also, the func itself will be a
bound one, which requires no 'self' argument.

This could have been written a little cleaner (inside the wrapper), but will
cost an overhead in runtime, which I try to avoid.